+++ source/Imports/ProcUtils.h
@@ -34,7 +34,7 @@ inline std::string ObterStr(uintptr_t address, int count) {
     uintptr_t hex[2] = { 0 };
     for (int i = 0; i < count; i++)
     {
-        classname = Ler<uintptr_t>(address + i * 4);
+        classname = Ler<uint32_t>(address + i * 4);
         hex[0] = (classname & 0xfffff000) >> 16;
         hex[1] = classname & 0xffff;
         buf16[m] = static_cast<UTF16>(hex[1]);
@@ -47,13 +47,13 @@ inline std::string ObterStr(uintptr_t address, int count) {
 }
 
 inline Vector3 Transform_ObterPosicao(uintptr_t Transform) {
-    uintptr_t TransformAcess = Ler<uintptr_t>(Transform + string2Offset(AY_OBFUSCATE("0x8")));
+    uintptr_t TransformAcess = Ler<uint32_t>(Transform + string2Offset(AY_OBFUSCATE("0x8")));
     int TransformIndex = Ler<int>(TransformAcess + string2Offset(AY_OBFUSCATE("0x24")));
-    uintptr_t TransformMatrix = Ler<uintptr_t>(TransformAcess + string2Offset(AY_OBFUSCATE("0x20")));
-    uintptr_t pTransformValues = Ler<uintptr_t>(TransformMatrix + string2Offset(AY_OBFUSCATE("0x18")));
+    uintptr_t TransformMatrix = Ler<uint32_t>(TransformAcess + string2Offset(AY_OBFUSCATE("0x20")));
+    uintptr_t pTransformValues = Ler<uint32_t>(TransformMatrix + string2Offset(AY_OBFUSCATE("0x18")));
     uintptr_t offsetPosition = string2Offset(AY_OBFUSCATE("0x30")) * static_cast<uintptr_t>(TransformIndex);
     Vector3 ResultPosition = Ler<Vector3>(pTransformValues + offsetPosition);
-    uintptr_t pOffsetCount = Ler<uintptr_t>(TransformMatrix + string2Offset(AY_OBFUSCATE("0x1C")));
+    uintptr_t pOffsetCount = Ler<uint32_t>(TransformMatrix + string2Offset(AY_OBFUSCATE("0x1C")));
     int IndexTransform = Ler<int>(pOffsetCount + string2Offset(AY_OBFUSCATE("0x4")) * TransformIndex);
     int curIndex = 0;
     while (IndexTransform >= 0) {
@@ -83,26 +83,26 @@ inline Vector3 Transform_ObterPosicao(uintptr_t Transform) {
 }
 
 inline void Transform_DefinirPosicao(uintptr_t Transform, Vector3 NovaPosicao) {
-    uintptr_t TransformAcess = Ler<uintptr_t>(Transform + string2Offset(AY_OBFUSCATE("0x8")));
+    uintptr_t TransformAcess = Ler<uint32_t>(Transform + string2Offset(AY_OBFUSCATE("0x8")));
     int TransformIndex = Ler<int>(TransformAcess + string2Offset(AY_OBFUSCATE("0x24")));
-    uintptr_t pTransformValues = Ler<uintptr_t>(Ler<uintptr_t>(TransformAcess + string2Offset(AY_OBFUSCATE("0x20"))) + string2Offset(AY_OBFUSCATE("0x18")));
+    uintptr_t pTransformValues = Ler<uint32_t>(Ler<uint32_t>(TransformAcess + string2Offset(AY_OBFUSCATE("0x20"))) + string2Offset(AY_OBFUSCATE("0x18")));
     uintptr_t offsetPosition = string2Offset(AY_OBFUSCATE("0x30")) * static_cast<uintptr_t>(TransformIndex);
     Escrever<Vector3>(pTransformValues + offsetPosition, NovaPosicao);
 }
 
 inline auto GetPlayerPosition = [](uintptr_t player, int positionType) -> Vector3 {
     if (positionType == 0) {
-        uintptr_t m_CachedTransform = Ler<uintptr_t>(player + Offsets::MainTransform);
+        uintptr_t m_CachedTransform = Ler<uint32_t>(player + Offsets::MainTransform);
         if (m_CachedTransform == 0) return Vector3{ 0, 0, 0 };
 
         return Transform_ObterPosicao(m_CachedTransform);
     }
 
     if (positionType == 1) {
-        uintptr_t OLCJOGDHJJJ5 = Ler<uintptr_t>(player + Offsets::MainTransform);
+        uintptr_t OLCJOGDHJJJ5 = Ler<uint32_t>(player + Offsets::MainTransform);
         if (OLCJOGDHJJJ5 == 0) return Vector3{ 0, 0, 0 };
 
-        uintptr_t TFNode = Ler<uintptr_t>(OLCJOGDHJJJ5 + string2Offset(AY_OBFUSCATE("0x8")));
+        uintptr_t TFNode = Ler<uint32_t>(OLCJOGDHJJJ5 + string2Offset(AY_OBFUSCATE("0x8")));
         if (TFNode == 0) return Vector3{ 0, 0, 0 };
 
         return Transform_ObterPosicao(TFNode);
@@ -111,22 +111,22 @@ inline auto GetPlayerPosition = [](uintptr_t player, int positionType) -> Vector
     };
 
 inline Vector3 ObterOssos(uintptr_t Player, uintptr_t Position) {
-    uintptr_t ListTransform = Ler<uintptr_t>(Player + Offsets::AIDDOCAPFKA);
+    uintptr_t ListTransform = Ler<uint32_t>(Player + Offsets::AIDDOCAPFKA);
     if (ListTransform != string2Offset(AY_OBFUSCATE("0")))
     {
-        uintptr_t Transform = Ler<uintptr_t>(ListTransform + string2Offset(AY_OBFUSCATE("0x8")));
+        uintptr_t Transform = Ler<uint32_t>(ListTransform + string2Offset(AY_OBFUSCATE("0x8")));
         if (Transform != string2Offset(AY_OBFUSCATE("0")))
         {
-            uintptr_t Location = Ler<uintptr_t>(Transform + Position);
+            uintptr_t Location = Ler<uint32_t>(Transform + Position);
             if (Location != string2Offset(AY_OBFUSCATE("0")))
             {
-                uintptr_t H1 = Ler<uintptr_t>(Location + string2Offset(AY_OBFUSCATE("0x8")));
+                uintptr_t H1 = Ler<uint32_t>(Location + string2Offset(AY_OBFUSCATE("0x8")));
                 if (H1 != string2Offset(AY_OBFUSCATE("0")))
                 {
-                    uintptr_t H2 = Ler<uintptr_t>(H1 + string2Offset(AY_OBFUSCATE("0x28")));
+                    uintptr_t H2 = Ler<uint32_t>(H1 + string2Offset(AY_OBFUSCATE("0x28")));
                     if (H2 != string2Offset(AY_OBFUSCATE("0")))
                     {
-                        uintptr_t H3 = Ler<uintptr_t>(H2 + string2Offset(AY_OBFUSCATE("0x14")));
+                        uintptr_t H3 = Ler<uint32_t>(H2 + string2Offset(AY_OBFUSCATE("0x14")));
                         if (H3 != string2Offset(AY_OBFUSCATE("0")))
                         {
                             uintptr_t H4 = H3 + string2Offset(AY_OBFUSCATE("0x60"));
@@ -144,32 +144,32 @@ inline Vector3 ObterOssos(uintptr_t Player, uintptr_t Position) {
 }
 
 inline Vector3 GetBonePositionV2(uintptr_t player, uintptr_t boneOffset) {
-    uintptr_t Location = Ler<uintptr_t>(player + boneOffset);
+    uintptr_t Location = Ler<uint32_t>(player + boneOffset);
     if (Location == 0 || Location < 0x10000) return Vector3{0, 0, 0};
-    uintptr_t H1 = Ler<uintptr_t>(Location + string2Offset(AY_OBFUSCATE("0x8")));
+    uintptr_t H1 = Ler<uint32_t>(Location + string2Offset(AY_OBFUSCATE("0x8")));
     if (H1 == 0 || H1 < 0x10000) return Vector3{0, 0, 0};
-    uintptr_t H2 = Ler<uintptr_t>(H1 + Offsets::Bones::Node_WorldPos);
+    uintptr_t H2 = Ler<uint32_t>(H1 + Offsets::Bones::Node_WorldPos);
     if (H2 == 0 || H2 < 0x10000) return Vector3{0, 0, 0};
-    uintptr_t H3 = Ler<uintptr_t>(H2 + string2Offset(AY_OBFUSCATE("0x14")));
+    uintptr_t H3 = Ler<uint32_t>(H2 + string2Offset(AY_OBFUSCATE("0x14")));
     if (H3 == 0 || H3 < 0x10000) return Vector3{0, 0, 0};
     return Ler<Vector3>(H3 + string2Offset(AY_OBFUSCATE("0x60")));
 }
 
 inline Vector3 GetHeadPosition(uintptr_t Entidade) {
-    uintptr_t ListTransform = Ler<uintptr_t>(Entidade + Offsets::AIDDOCAPFKA);
+    uintptr_t ListTransform = Ler<uint32_t>(Entidade + Offsets::AIDDOCAPFKA);
     if (ListTransform > 0x10000) {
         bool Garota = Ler<bool>(Entidade + Offsets::CDOBMFNCJHD);
         uintptr_t boneOffset = Garota ? string2Offset(AY_OBFUSCATE("0x3C")) : string2Offset(AY_OBFUSCATE("0x38"));
 
-        uintptr_t Transform = Ler<uintptr_t>(ListTransform + string2Offset(AY_OBFUSCATE("0x8")));
+        uintptr_t Transform = Ler<uint32_t>(ListTransform + string2Offset(AY_OBFUSCATE("0x8")));
         if (Transform > 0x10000) {
-            uintptr_t Location = Ler<uintptr_t>(Transform + boneOffset);
+            uintptr_t Location = Ler<uint32_t>(Transform + boneOffset);
             if (Location > 0x10000) {
-                uintptr_t H1 = Ler<uintptr_t>(Location + string2Offset(AY_OBFUSCATE("0x8")));
+                uintptr_t H1 = Ler<uint32_t>(Location + string2Offset(AY_OBFUSCATE("0x8")));
                 if (H1 > 0x10000) {
-                    uintptr_t H2 = Ler<uintptr_t>(H1 + string2Offset(AY_OBFUSCATE("0x28")));
+                    uintptr_t H2 = Ler<uint32_t>(H1 + string2Offset(AY_OBFUSCATE("0x28")));
                     if (H2 > 0x10000) {
-                        uintptr_t H3 = Ler<uintptr_t>(H2 + string2Offset(AY_OBFUSCATE("0x14")));
+                        uintptr_t H3 = Ler<uint32_t>(H2 + string2Offset(AY_OBFUSCATE("0x14")));
                         if (H3 > 0x10000) {
                             Vector3 headPos = Ler<Vector3>(H3 + string2Offset(AY_OBFUSCATE("0x60")));
                             if (headPos.X != 0 || headPos.Y != 0 || headPos.Z != 0) {
@@ -182,7 +182,7 @@ inline Vector3 GetHeadPosition(uintptr_t Entidade) {
         }
     }
 
-    uintptr_t mainTransform = Ler<uintptr_t>(Entidade + Offsets::MainTransform);
+    uintptr_t mainTransform = Ler<uint32_t>(Entidade + Offsets::MainTransform);
     if (mainTransform > 0x10000) {
         return Transform_ObterPosicao(mainTransform);
     }
